requirements.txt
requests
python-dotenv
pyyaml
tenacity
pandas
pytest